requirements.txt
Django>=5.2,<5.3
Pillow>=10.0
python-dotenv>=1.0
requests>=2.31
